PIM-3725: extract getNonEligibleAttributes from ProductTemplateBuilder to a dedicated service
