+++ spec/Pim/Bundle/CatalogBundle/Builder/ProductTemplateBuilderSpec.php
@@ -6,10 +6,8 @@ use Doctrine\Common\Collections\Collection;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Builder\ProductBuilder;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductTemplateInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
-use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
@@ -19,14 +17,12 @@ class ProductTemplateBuilderSpec extends ObjectBehavior
     function let(
         NormalizerInterface $normalizer,
         DenormalizerInterface $denormalizer,
-        AttributeRepositoryInterface $attributeRepository,
         ProductBuilder $productBuilder
     ) {
         $this->beConstructedWith(
             $normalizer,
             $denormalizer,
             $productBuilder,
-            $attributeRepository,
             'Pim\Bundle\CatalogBundle\Entity\ProductTemplate',
             'Pim\Bundle\CatalogBundle\Model\Product'
         );
@@ -40,30 +36,6 @@ class ProductTemplateBuilderSpec extends ObjectBehavior
     function it_creates_product_templates()
     {
         $this->createProductTemplate()->shouldReturnAnInstanceOf('Pim\Bundle\CatalogBundle\Entity\ProductTemplate');
-    }
-
-    function it_returns_non_eligible_attributes(
-        $attributeRepository,
-        GroupInterface $group,
-        ProductTemplateInterface $template,
-        AttributeInterface $length,
-        AttributeInterface $name,
-        AttributeInterface $color,
-        AttributeInterface $identifier,
-        Collection $collection
-    ) {
-        $group->getProductTemplate()->willReturn($template);
-        $group->getAxisAttributes()->willReturn($collection);
-        $collection->toArray()->willReturn([$length]);
-
-        $template->getValuesData()->willReturn(['name' => 'foo', 'color' => 'bar']);
-
-        $attributeRepository->findOneByIdentifier('name')->willReturn($name);
-        $attributeRepository->findOneByIdentifier('color')->willReturn($color);
-        $attributeRepository->findBy(['unique' => true])->willReturn([$name, $identifier]);
-
-        $attributes = [$length, $name, $color, $identifier];
-        $this->getNonEligibleAttributes($group)->shouldReturn($attributes);
     }
 
     function it_adds_attributes_to_a_product_template(
