+++ src/Pim/Bundle/CatalogBundle/Resources/config/builders.yml
@@ -17,6 +17,5 @@ services:
             - '@pim_serializer'
             - '@pim_serializer'
             - '@pim_catalog.builder.product'
-            - '@pim_catalog.repository.attribute'
             - %pim_catalog.entity.product_template.class%
             - %pim_catalog.entity.product.class%
