+++ src/Pim/Bundle/EnrichBundle/Controller/VariantGroupController.php
@@ -8,6 +8,7 @@ use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\CatalogBundle\Factory\GroupFactory;
 use Pim\Bundle\CatalogBundle\Manager\GroupManager;
 use Pim\Bundle\CatalogBundle\Manager\ProductTemplateBuilder;
+use Pim\Bundle\CatalogBundle\Manager\VariantGroupAttributesResolver;
 use Pim\Bundle\CatalogBundle\Model\AvailableAttributes;
 use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
@@ -38,23 +39,27 @@ class VariantGroupController extends GroupController
     protected $attributeRepository;
 
     /** @var ProductTemplateBuilder */
-    protected $tplAttributesManager;
+    protected $templateBuilder;
+
+    /** @var VariantGroupAttributesResolver */
+    protected $groupAttributesResolver;
 
     /**
-     * @param Request                      $request
-     * @param EngineInterface              $templating
-     * @param RouterInterface              $router
-     * @param SecurityContextInterface     $securityContext
-     * @param FormFactoryInterface         $formFactory
-     * @param ValidatorInterface           $validator
-     * @param TranslatorInterface          $translator
-     * @param EventDispatcherInterface     $eventDispatcher
-     * @param GroupManager                 $groupManager
-     * @param HandlerInterface             $groupHandler
-     * @param Form                         $groupForm
-     * @param GroupFactory                 $groupFactory
-     * @param AttributeRepositoryInterface $attributeRepository
-     * @param ProductTemplateBuilder       $tplAttributesManager
+     * @param Request                        $request
+     * @param EngineInterface                $templating
+     * @param RouterInterface                $router
+     * @param SecurityContextInterface       $securityContext
+     * @param FormFactoryInterface           $formFactory
+     * @param ValidatorInterface             $validator
+     * @param TranslatorInterface            $translator
+     * @param EventDispatcherInterface       $eventDispatcher
+     * @param GroupManager                   $groupManager
+     * @param HandlerInterface               $groupHandler
+     * @param Form                           $groupForm
+     * @param GroupFactory                   $groupFactory
+     * @param AttributeRepositoryInterface   $attributeRepository
+     * @param ProductTemplateBuilder         $templateBuilder
+     * @param VariantGroupAttributesResolver $groupAttributesResolver
      */
     public function __construct(
         Request $request,
@@ -70,7 +75,8 @@ class VariantGroupController extends GroupController
         Form $groupForm,
         GroupFactory $groupFactory,
         AttributeRepositoryInterface $attributeRepository,
-        ProductTemplateBuilderInterface $templateBuilder
+        ProductTemplateBuilderInterface $templateBuilder,
+        VariantGroupAttributesResolver $groupAttributesResolver
     ) {
         parent::__construct(
             $request,
@@ -89,6 +95,7 @@ class VariantGroupController extends GroupController
 
         $this->attributeRepository = $attributeRepository;
         $this->templateBuilder = $templateBuilder;
+        $this->groupAttributesResolver = $groupAttributesResolver;
     }
 
     /**
@@ -175,7 +182,7 @@ class VariantGroupController extends GroupController
         return $this->createForm(
             'pim_available_attributes',
             new AvailableAttributes(),
-            ['excluded_attributes' => $this->templateBuilder->getNonEligibleAttributes($group)]
+            ['excluded_attributes' => $this->groupAttributesResolver->getNonEligibleAttributes($group)]
         );
     }
 }
