+++ src/Pim/Bundle/EnrichBundle/Resources/config/controllers.yml
@@ -190,6 +190,7 @@ services:
             - '@pim_catalog.factory.group'
             - '@pim_catalog.repository.attribute'
             - '@pim_catalog.builder.product_template'
+            - '@pim_catalog.resolver.variant_group_attributes'
 
     pim_enrich.controller.group_type:
         scope: request
@@ -248,3 +249,4 @@ services:
             - '@pim_catalog.saver.group'
             - '@pim_catalog.repository.attribute'
             - '@pim_catalog.builder.product_template'
+            - '@pim_catalog.resolver.variant_group_attributes'
